Try to get the file containing main.main from line number data
We now attempt to determine the file containing main.main by finding
the function in the executable, recovering its start address, and then
mapping this address through debug/gosym.Table.PCToLine(), which if it
works will reliably give us the file. If we can determine the file
this way, we then attempt to turn it into a package name by checking
if GOROOT or any GOPATH directory is a prefix of it. If any of this
fails, we fall back to various degrees of the original heuristics.

This improves Import's reliability on executables in various
circumstances.

This also includes a bugfix, replacing strings.TrimRight() with the
intended strings.TrimSuffix().